cp-17: * move auth error exception to shared module
