+++ backend/src/libs/exceptions/exceptions.ts
@@ -1,2 +1,1 @@
-export { AuthError } from './auth-error/auth-error.exception.js';
-export { ValidationError } from 'shared/build/index.js';
+export { AuthError, ValidationError } from 'shared/build/index.js';
