+++ shared/src/libs/exceptions/exceptions.ts
@@ -1,3 +1,4 @@
 export { ApplicationError } from './application-error/application-error.exception.js';
+export { AuthError } from './auth-error/auth-error.exception.js';
 export { HTTPError } from './http-error/http-error.exception.js';
 export { ValidationError } from './validation-error/validation-error.exception.js';
